rename landing page folder
