+++ components/landingPage/EmailCapture.tsx
