+++ components/landingPage/TermsAndConditions.tsx
